BlockTransactionSelector refactoring (#5931)
