+++ plugin-api/src/main/java/org/hyperledger/besu/plugin/services/txselection/TransactionSelector.java
@@ -17,28 +17,17 @@ package org.hyperledger.besu.plugin.services.txselection;
 
 import org.hyperledger.besu.datatypes.Transaction;
 import org.hyperledger.besu.plugin.Unstable;
-import org.hyperledger.besu.plugin.data.Log;
 import org.hyperledger.besu.plugin.data.TransactionSelectionResult;
-
-import java.util.List;
 
 /** Interface for the transaction selector */
 @Unstable
 public interface TransactionSelector {
-
   /**
-   * Method called to decide whether a transaction is added to a block. The method can also indicate
+   * Method called to decide whether a transaction is added to a block. The result can also indicate
    * that no further transactions can be added to the block.
    *
    * @param transaction candidate transaction
-   * @param success true, if the transaction executed successfully
-   * @param logs the logs created by this transaction
-   * @param cumulativeGasUsed gas used by this and all previous transaction in the block
    * @return TransactionSelectionResult that indicates whether to include the transaction
    */
-  TransactionSelectionResult selectTransaction(
-      final Transaction transaction,
-      final boolean success,
-      final List<Log> logs,
-      final long cumulativeGasUsed);
+  TransactionSelectionResult evaluateTransactionPreProcessing(Transaction transaction);
 }
